added the EngineThread
